Changed all_data dictionary to have sensibly named keys, static view page, protocol names to protocl entry in database
Unfortunately this breaks the "edit" view of form sheets, I'm trying to fix that.

The main this is that renaming the keys in the all_data dict makes them accessible in django templates which I think is an important change moving forward.

Also, the name of each project shouldn't be a default auto-increment value. They should be pulled from the all_data at time of entry update so that it can be displayed on the index page. Client said as much last meeting.