Made changes to cover sheet form
Implemented changes as per client requests from last meeting to the cover sheet form. Changed "type of sheet" to a checkbox selection as per one of the examples, and made actions and interventions editable (though not add/delete boxes, could use "additional" field for now for that). Also consider in the future the ability to upload images or other files?